server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = "Currencies servive"
authors = ["BogPelmeney"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.6"
djangorestframework = "^3.13.1"
requests = "^2.28.1"
beautifulsoup4 = "^4.11.1"
openpyxl = "^3.0.10"
django-cors-headers = "^3.13.0"
django-filter = "^22.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
